fix(evidence): object_store 0.13 compatibility (ObjectStoreExt, bump to 0.13)
- Add ObjectStoreExt to scope so get/head/put resolve (0.13 API)
- Bump object_store from 0.11 to 0.13 in assay-evidence

Fixes Smoke Install (E2E) failure for object_store 0.13.1.
Supersedes Dependabot PR #1075.

+++ crates/assay-evidence/src/store/object_store_backend.rs
@@ -7,7 +7,7 @@ use std::sync::Arc;
 use async_trait::async_trait;
 use bytes::Bytes;
 use futures::TryStreamExt;
-use object_store::{ObjectStore, PutMode, PutOptions, PutPayload};
+use object_store::{ObjectStore, ObjectStoreExt, PutMode, PutOptions, PutPayload};
 
 use super::{BundleMeta, BundleStore, KeyBuilder, StoreError, StoreResult, StoreSpec};
 
@@ -305,7 +305,7 @@ impl BundleStore for ObjectStoreBundleStore {
                     .parse_bundle_key(&entry.location)
                     .map(|id| BundleMeta {
                         bundle_id: id,
-                        size: Some(entry.size as u64),
+                        size: Some(entry.size),
                         modified: Some(entry.last_modified),
                     })
             })
